feat: feedback (#66)
* feat: added javadoc in utils, implementation of FeedbackCalculator, fixed names on method/class

* feat: added 40 new cards (20 medium, 20 hard)

* feat: fixxed values for feedbackEvaluator

* feat: Changed co2 datatype into double (kg)

* feat: updated co2 value

---------

+++ src/main/java/com/ecologicstudios/utils/Choice.java
@@ -1,20 +1,51 @@
+/**
+ * The {@code Choice} class represents an alternative option for a card scenario,
+ * including its descriptive text and associated CO2 value.
+ * <p>
+ * Each {@code Choice} object contains the choice description and the CO2 impact
+ * of selecting that choice.
+ *
+ * @author EcoLogic Studios
+ */
 package com.ecologicstudios.utils;
 
 public class Choice{
+    /**
+     * The description of the choice presented to the player.
+     */
     String choice;
-    private int co2;
+    /**
+     * The CO2 value associated with this choice.
+     */    
+    private double co2;
 
-    public Choice(String text, int co2) {
+    /**
+     * Constructs a Choice with the specified description and CO2 value.
+     *
+     * @param text the description of the choice
+     * @param co2 the CO2 value associated with the choice
+     */
+    public Choice(String text, double co2) {
         this.choice = text;
 
         this.co2 = co2;
     }
 
+    /**
+     * Returns the description of this choice.
+     *
+     * @return the choice description
+     */
     public String getChoice(){
         return choice;
     }
 
-    public int getco2(){
+    /**
+    * Returns the CO2 value associated with this choice.
+    *
+    * @return the CO2 value
+    */
+    public double getco2(){
         return co2;
     }
 }
